enable contextual spans
Previously, tracing-actions required explicit parent spans. That's
nice and is faster when you have it available, but sometimes the
resulting code is rough. This change adds support for tracing's
notion of is_contextual() and, when you haven't specified an explicit
parent, it tries to wire a new span to the current span.