feat: add a should_record_span predicate to ActionTraceSubscriber
A user may want some way to filter out recorded spans before sending
them to a `Sink`. How the user wants to determine whether an
`ActionSpan` is recorded is up to them, but this PR gives them the
ability to do such.

The simplest approach is to perform the check before we grab the mutex
and send the `ActionSpan` to the `Sink`. If the predicate returns false,
we avoid spending CPU cycles grabbing the lock and appending the span to
the end of the `Vec`. It then gets dropped from memory regardless of the
outcome (as it always has).